- compose 배터리 보류 => xml로 진행

+++ app/src/main/java/com/example/qwer_test_kt/domin/model/BatteryInfo.kt
@@ -1,6 +1,6 @@
 package com.example.qwer_test_kt.domin.model
 
 data class BatteryInfo(
-    val percentage: Float,
+    val percentage: Int,
     val status: Int
 )
